+++ chapter3/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "dist/index.js",
   "scripts": {
-    "start:dev": "nodemon index.ts",
+    "start:dev": "nodemon index.js",
     "build": "rimraf ./build && tsc",
     "start": "npm run build && node build/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
