+++ package.json
@@ -7,6 +7,7 @@
     "@types/express": "^4.17.8",
     "@types/node": "^14.6.2",
     "express": "^4.17.1",
+    "node-readability": "^3.0.0",
     "typescript": "^4.0.2"
   },
   "devDependencies": {
